Broke out fetch logic into utils.tsx Renamed some function to clearify what they do
