+++ common/utils.ts
@@ -1,4 +1,4 @@
-import { WorkItem } from './types'
+import { WorkItem, ExperienceItem } from './types'
 import { db } from '@/lib/firebaseAdmin'
 import { notFound } from 'next/navigation'
 
@@ -13,7 +13,7 @@ export const fetchWork = async () => {
 }
 
 export const fetchSingleWork = async (slug: string) => {
-    const snapshot = await db
+    const singleWorkSnapshot = await db
         .collection('work')
         .where('slug', '==', slug)
         .limit(1)
@@ -22,11 +22,38 @@ export const fetchSingleWork = async (slug: string) => {
     // console.log(page)
     // Check cover, if missing return null
 
-    if (snapshot.empty) {
+    if (singleWorkSnapshot.empty) {
         console.error('No page found for this slug:', slug)
         notFound() // Show Next.js 404
     }
 
-    const page = snapshot.docs.map((doc) => doc.data())[0]
+    const page = singleWorkSnapshot.docs.map((doc) => doc.data())[0]
     return page
+}
+
+export const fetchExperience = async () => {
+    // Firebase way of fetching data for "collections" of data
+    const experienceSnapshot = await db
+        .collection('experience')
+        .orderBy('end', 'desc')
+        .get()
+    const experienceData: ExperienceItem[] = experienceSnapshot.docs.map(
+        (doc) => ({
+            id: doc.id,
+            ...(doc.data() as Omit<ExperienceItem, 'id'>), // assuming id comes from doc.id
+        })
+    )
+
+    return experienceData
+}
+
+export const fetchEducation = async () => {
+    const eduationSnapshot = await db.collection('education').get()
+    const educationData: ExperienceItem[] = eduationSnapshot.docs.map(
+        (doc) => ({
+            id: doc.id,
+            ...(doc.data() as Omit<ExperienceItem, 'id'>), // assuming id comes from doc.id
+        })
+    )
+    return educationData
 }
